backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.11
dj-database-url==3.1.2
python-dotenv==1.0.0
openai==1.3.7
Pillow==10.1.0
